requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow-cpu==2.20.0
numpy==1.24.3
Pillow==10.0.0
Werkzeug==2.3.7
